backend/requirements.txt
langgraph>=0.2.0
langchain-ollama>=0.1.0
langchain-core>=0.2.0
# duckduckgo-search>=6.0.0
python-dotenv==1.2.2
langchain-groq==1.1.2
fastapi[standard]
